Implement Images per bin
Removed imagepaths state from controller, this is now handled by a getpaths directly from the db.
Controller does not need to have a local state to the paths, this should be on the grids level, and each grid keeps a state of what paths they contain.
Implement database to have multiple bins and return paths based on bin id.