feat: Balance and export
Balance dataset for ML Training.
Balance & Export creates a ballanced copy of the dataset with images split into training, validation and test sets.
Dataset is balanced based on the number of images in the smallest bin with a 60, 20, 20 split.
Images are never selected twice.
Images are selected randomly to go into training, validation and test sets.